requirements.txt
python-pptx>=0.6.23
python-docx>=1.1.0
openpyxl>=3.1.2
